+++ release_notes.md
@@ -2,6 +2,7 @@
   - [#257](https://github.com/Mehdik/Humanizer/pull/257): Added German localisation for ToOrdinalWords and Ordinalize
   - [#261](https://github.com/Mehdik/Humanizer/pull/261): Added future dates to Portuguese - Brazil
   - [#269](https://github.com/MehdiK/Humanizer/pull/269): Added Vietnamese localisation
+  - [#268](https://github.com/Mehdik/Humanizer/pull/268): Added humanization of collections
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.25.4...master)